core: fixing domain referencing into master data client core: implement a running first version of insights collector service
