+++ src/insights/collector/WeatherInsights.Collector.slnx
@@ -1,0 +1,13 @@
+<Solution>
+  <Folder Name="/Core/">
+    <Project Path="Application/WeatherInsights.Collector.Application.csproj" />
+    <Project Path="Domain/WeatherInsights.Collector.Domain.csproj" />
+  </Folder>
+  <Folder Name="/Infrastructure/">
+    <Project Path="Infrastructure/WeatherInsights.Collector.Infrastructure.csproj" />
+  </Folder>
+  <Folder Name="/Presentation/">
+    <Project Path="Presentation/WfInsights.Collector.Api.csproj" />
+  </Folder>
+  <Folder Name="/Solution Files/" />
+</Solution>
