+++ src/master-data/service/Api/Controllers/CitiesController.cs
@@ -1,3 +1,4 @@
+using MasterData.Api.Factories;
 using MasterData.Application.Services.Interfaces;
 using MasterData.Client.Dtos.Responses.City.GetAll;
 using MasterData.Domain.AggregateModel.Cities;
