+++ src/master-data/service/Api/Program.cs
@@ -24,14 +24,6 @@ var app = builder.Build();
 
 app.UseSwaggerSwashbuckle();
 
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
 app.UseHttpsRedirection();
 
 app.UseAuthorization();
