+++ src/master-data/service/Client/Dtos/Responses/City/GetAll/GpsCoordinatesDto.cs
@@ -1,22 +1,3 @@
-using MasterData.Domain.AggregateModel.Cities.ValueObjects;
-
 namespace MasterData.Client.Dtos.Responses.City.GetAll;
 
 public record GpsCoordinatesDto(decimal Latitude, decimal Longitude);
-
-
-/// <summary>
-/// Factory for <see cref="GpsCoordinatesDto"/>
-/// </summary>
-public static class GpsCoordinatesDtoFactory
-{
-    /// <summary>
-    /// Creates a dto object from a domain <see cref="GpsCoordinates"/> entity
-    /// </summary>
-    /// <param name="coordinate"></param>
-    /// <returns></returns>
-    public static GpsCoordinatesDto 
-        CreateFromDomain(GpsCoordinates coordinate) =>
-        new GpsCoordinatesDto(coordinate.Latitude, coordinate.Longitude);
-    
-}
